python/mysql-demo/requirements.txt
mysql-connector-python>=8.0.24